Fix esp_idf build.

+++ components/truma_inetbox/binary_sensor/TrumaHeaterBinarySensor.cpp
@@ -38,7 +38,7 @@ void TrumaHeaterBinarySensor::setup() {
 
 void TrumaHeaterBinarySensor::dump_config() {
   ESP_LOGCONFIG("", "Truma Heater Binary Sensor");
-  ESP_LOGCONFIG(TAG, "Type %u", this->type_);
+  ESP_LOGCONFIG(TAG, "Type %u", (size_t) this->type_);
 }
 }  // namespace truma_inetbox
 }  // namespace esphome
